data: cc-bootchart.service
This patch adds a new service to copy bootchart svg to the
container rootfs.

The cc-agent.target Wants this new service, so in case
the service is not part of the image wont be executed.

This allow run docker and get the bootchart from the VM:

docker run -ti clearlinux bash -c 'sleep 3; cat /run/*.svg' >
bootchart.svg

The kernel must boot with systemd-bootchart to generate a chart,
this can be added in data/vm.json .

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>