configure: Dont fail build with missing m4 files
We request to to have m4 files to do extra valgrind and code coverage
checks:
- ax_valgrind_check.html
- ax_code_coverage.m4

If the this files are not present the build should not fail.

This patch modifies MACRO_CHECK macro is a definition does not exist.
Now MACRO_CHECK defines a variable HAS_${MACRO} this can be used
to know if the MACRO can be used or not. This will allow to no
fail a build.

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>